rename ResponsiveFooter to Footer, rename ResponsiveHeader to header. include login link to the menu
