+++ src/components/Footer/FooterTablet.tsx
