+++ src/components/Header/components/SearchBarMobile.tsx
