+++ src/components/Header/components/SearchBarTablet.tsx
