+++ src/layouts/index.tsx
@@ -3,8 +3,8 @@ import PropTypes from 'prop-types';
 import { Outlet } from 'react-router-dom';
 import { Container, styled } from '@mui/material';
 import { Helmet } from 'react-helmet-async';
-import Footer from '../components/ResponsiveFooter';
-import Header from '../components/ResponsiveHeader';
+import Footer from '../components/Footer';
+import Header from '../components/Header';
 import ScrollButton from '../components/ScrollButton/ScrollButtonView';
 
 const StyledContainer = styled(Container)(() => ({
